Separate unnecessary return none mixin
